Shift object freezing into a trait under Component.
Will maybe move it, plus locking, to another composer lib.
