+++ core/tests/Drupal/Tests/Core/Asset/Collection/AssetLibraryTest.php
@@ -7,7 +7,7 @@
 namespace Drupal\Tests\Core\Asset\Collection;
 
 use Drupal\Core\Asset\Collection\AssetLibrary;
-use Drupal\Core\Asset\Exception\FrozenObjectException;
+use Drupal\Component\ObjectState\FrozenObjectException;
 use Drupal\Tests\Core\Asset\AssetUnitTest;
 
 /**
